Refactor Payload class to Configuration
