+++ ipassign/__init__.py
@@ -1,9 +1,9 @@
 from .protocol.enums import acknowledgements, commands
 from .protocol.message import MAX_PACKET_LENGTH, Message, MIN_PACKET_LENGTH
-from .protocol.payload import Acknowledgement, Payload
+from .protocol.payload import Acknowledgement, Configuration
 
 from .networking import is_known_hostname, MULTICAST_ADDR, MULTICAST_PORT
 
-__all__ = [acknowledgements, Acknowledgement, commands, is_known_hostname,
-           MAX_PACKET_LENGTH, Message, MIN_PACKET_LENGTH, MULTICAST_ADDR,
-           MULTICAST_PORT, Payload]
+__all__ = [acknowledgements, Acknowledgement, commands, Configuration,
+           is_known_hostname, MAX_PACKET_LENGTH, Message, MIN_PACKET_LENGTH,
+           MULTICAST_ADDR, MULTICAST_PORT]
