+++ tests/test_data.py
@@ -1,7 +1,8 @@
 # PACKET is a discovery packet, see README
 PACKET = b'\x78\x45\xc4\xf7\x8f\x48\x00\x00\x01\x00\x02\x00\x00\x00\x31\x8f\x64\x48'  # noqa
 
-# REPLY is a reply to the above discovery request, an icepap describing its configuration
+# REPLY is a reply to the above discovery request,
+# an icepap describing its configuration
 REPLY = b'\x00\x0c\xc6\x69\x13\x2d\x01\x00\x00\x00\x03\x00\x38\x00\x00\x22\x19\x06\xbf\x58\x00\x0c\xc6\x69\x13\x2d\xac\x18\x9b\xde\xac\x18\x9b\xff\xff\xff\xff\x00\xac\x18\x9b\x63\x00\x0c\xc6\x69\x13\x2d\x00\x00\x00\x00\x69\x63\x65\x65\x75\x34\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xb3\x57\x23\x0d'  # noqa
 
 # PAYLOAD is the payload from REPLY, contains the data described in the README
